refactor(MoviesAdapter): 뷰의 값을 세팅하는 역할을 ViewHolder로 이동
